views fixes
