+++ devinsight/lib/main.dart
@@ -9,7 +9,6 @@ void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await StorageService.init();
   //await dotenv.load(fileName: '.env');
-
   runApp(
     const ProviderScope(
       child: MyApp(),
