+++ devinsight/lib/ui/home/views/profile.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class Profile extends StatelessWidget {
+  const Profile({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Profile'),
+    );
+  }
+}
